Move cobra_* into agent directory
The cobra_manager and cobra_client are only used inside the agent,
therefore we move them into the agent directory, right next to
aci_agent.py, which is using them.
